files dir changed for curriculum
