+++ apps/api/src/modules/faculty-activity/curriculum/curriculum.controller.spec.ts
