+++ apps/api/src/modules/faculty-activity/curriculum/curriculum.controller.ts
@@ -1,5 +1,5 @@
 import { GlobalExceptionFilter } from '@api/exceptions/exception-filter';
-import { CourseCurriculumService } from '@api/modules/faculty-activity/curriculum/curriculum/curriculum.service';
+import { CourseCurriculumService } from '@api/modules/faculty-activity/curriculum/curriculum.service';
 import {
   Body,
   Controller,
