+++ apps/api/src/modules/faculty-activity/curriculum/curriculum.service.ts
