--- apps/api/src/modules/faculty-activity/curriculum/curriculum/curriculum.module.ts
@@ -1,13 +1,0 @@
-import { Module } from '@nestjs/common';
-import { PrismaService } from '@api/modules/persistance/prisma/prisma.service';
-import { CourseController } from '@api/modules/faculty-activity/course/course.controller';
-import { CourseCurriculumService } from '@api/modules/faculty-activity/curriculum/curriculum/curriculum.service';
-import { CurriculumController } from '@api/modules/faculty-activity/curriculum/curriculum/curriculum.controller';
-
-@Module({
-  imports: [],
-  providers: [CourseCurriculumService, PrismaService],
-  controllers: [CurriculumController],
-  exports: [],
-})
-export class CurriculumModule {}
